Working on the for loop
I'm trying to get the td.textcontent to get the right values from the
objects that are in the array. Even if key value pairs in an object had
indexes they still wouldn't line up with the columns in my table because
I'm only using 3 of the key value pairs. Hmmm... Break time.